Feat: Animation 별도로 분리
